Move SAP_Internal_Stats_KeyFigure extension to its own file, group some files
